Refactor content into own util
Having the yaml content in the data.js utility was creating problems
with Jest. In order to avoid them, I refactored the content helper into
its own file
