+++ src/components/Explorer.js
@@ -14,7 +14,8 @@ import { fetchSummaries } from '../actions/summaryActions'
 import { fetchNibrsDimensions } from '../actions/nibrsActions'
 import { updateFilters, updateFiltersAndUrl } from '../actions/filterActions'
 import { hideSidebar, showSidebar } from '../actions/sidebarActions'
-import { content, crimeTypes } from '../util/data'
+import { crimeTypes } from '../util/data'
+import content from '../util/content'
 import { slugify } from '../util/text'
 import lookup from '../util/usa'
